fix ICS03 proofs
