+++ x/ibc/02-client/client/cli/cli.go
@@ -1,9 +1,10 @@
 package cli
 
 import (
+	"github.com/spf13/cobra"
+
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/spf13/cobra"
 )
 
 // GetQueryCmd returns the query commands for IBC clients
