+++ x/ibc/03-connection/client/rest/query.go
@@ -5,11 +5,10 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
-	abci "github.com/tendermint/tendermint/abci/types"
 
 	"github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/types/rest"
-	"github.com/cosmos/cosmos-sdk/x/ibc/03-connection/types"
+	"github.com/cosmos/cosmos-sdk/x/ibc/03-connection/client/utils"
 )
 
 func registerQueryRoutes(cliCtx context.CLIContext, r *mux.Router, queryRoute string) {
@@ -32,30 +31,21 @@ func queryConnectionHandlerFn(cliCtx context.CLIContext, queryRoute string) http
 	return func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
 		connectionID := vars[RestConnectionID]
+		prove := rest.ParseQueryProve(r)
 
 		cliCtx, ok := rest.ParseQueryHeightOrReturnBadRequest(w, cliCtx, r)
 		if !ok {
 			return
 		}
 
-		req := abci.RequestQuery{
-			Path:  "store/ibc/key",
-			Data:  types.KeyConnection(connectionID),
-			Prove: rest.ParseQueryProve(r),
-		}
-
-		res, err := cliCtx.QueryABCI(req)
+		connRes, err := utils.QueryConnection(cliCtx, connectionID, prove)
 		if err != nil {
 			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
 			return
 		}
 
-		var connection types.ConnectionEnd
-		if err := cliCtx.Codec.UnmarshalBinaryLengthPrefixed(res.Value, &connection); err != nil {
-			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
-		}
-		cliCtx = cliCtx.WithHeight(res.Height)
-		rest.PostProcessResponse(w, cliCtx, types.NewConnectionResponse(connectionID, connection, res.Proof, res.Height))
+		cliCtx = cliCtx.WithHeight(int64(connRes.ProofHeight))
+		rest.PostProcessResponse(w, cliCtx, connRes)
 	}
 }
 
@@ -74,29 +64,20 @@ func queryClientConnectionsHandlerFn(cliCtx context.CLIContext, queryRoute strin
 	return func(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
 		clientID := vars[RestClientID]
+		prove := rest.ParseQueryProve(r)
 
 		cliCtx, ok := rest.ParseQueryHeightOrReturnBadRequest(w, cliCtx, r)
 		if !ok {
 			return
 		}
 
-		req := abci.RequestQuery{
-			Path:  "store/ibc/key",
-			Data:  types.KeyClientConnections(clientID),
-			Prove: rest.ParseQueryProve(r),
-		}
-
-		res, err := cliCtx.QueryABCI(req)
+		connPathsRes, err := utils.QueryClientConnections(cliCtx, clientID, prove)
 		if err != nil {
 			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
 			return
 		}
 
-		var paths []string
-		if err := cliCtx.Codec.UnmarshalBinaryLengthPrefixed(res.Value, &paths); err != nil {
-			rest.WriteErrorResponse(w, http.StatusInternalServerError, err.Error())
-		}
-		cliCtx = cliCtx.WithHeight(res.Height)
-		rest.PostProcessResponse(w, cliCtx, types.NewClientConnectionsResponse(clientID, paths, res.Proof, res.Height))
+		cliCtx = cliCtx.WithHeight(int64(connPathsRes.ProofHeight))
+		rest.PostProcessResponse(w, cliCtx, connPathsRes)
 	}
 }
